% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{lreg5-class}
\alias{lreg5-class}
\title{An S4 class to Logistic Regression.}
\description{
An S4 class to Logistic Regression.
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{Coefficients}

\item{\code{var}}{Variance Covariance Matrix}

\item{\code{deviance}}{Deviance}

\item{\code{predictors}}{Predictors of the model}

\item{\code{iterations}}{No of iterations for convergence}
}}

