% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lreg5}
\alias{lreg5}
\title{carry out regression}
\usage{
lreg5(X, y, predictors = colnames(X), max.iter = 10, tol = 1e-06,
  constant = TRUE, ...)
}
\arguments{
\item{X,y}{inputs}

\item{predictors}{defaults to \code{colnames(x)}}

\item{max.iter}{maximum number of iterations (default: 10)}

\item{tol}{tolerance}

\item{constant}{add a constant}

\item{...}{Additional arguements (currently ignored)}
}
\value{
an object of class lreg5
}
\description{
carry out regression
}
\examples{
lreg5(X=matrix(rnorm(3*100),ncol=3, dimnames = list(NULL, letters[1:3])),
      y=sample(0:1,100, replace=TRUE))
}
\seealso{
\code{\link{lreg5-class}}
}

